+++ src/database/models/capability/mod.rs
@@ -7,7 +7,7 @@ pub use entity::*;
 pub use forms::*;
 
 /// All the capability of the application
-pub const CAPABILITIES: [&str; 8] = [
+pub const CAPABILITIES: [&str; 9] = [
     "role:manage",
     "user:manage_role",
     "users:view",
@@ -16,4 +16,5 @@ pub const CAPABILITIES: [&str; 8] = [
     "post:update",
     "post:delete",
     "post:view_hidden",
+    "post:edit_locked",
 ];